Added basic screen setup
