--- core/src/com/spaulding/salt/Screen/MenuScreen.java
@@ -1,32 +1,0 @@
-package com.spaulding.salt.Screen;
-
-import com.badlogic.gdx.scenes.scene2d.Stage;
-
-/**
- * Created by jared on 5/12/2016.
- */
-public class MenuScreen extends Screen {
-    public void create() {
-
-    }
-
-    public void render(Stage stage) {
-
-    }
-
-    public void resize(int width, int height) {
-
-    }
-
-    public void dispose() {
-
-    }
-
-    public void pause() {
-
-    }
-
-    public void resume() {
-
-    }
-}
