--- core/src/com/spaulding/salt/Screen/ScreenManager.java
@@ -1,23 +1,0 @@
-package com.spaulding.salt.Screen;
-
-/**
- * Created by jared on 5/12/2016.
- */
-public class ScreenManager {
-
-    private static Screen current_screen;
-
-    public static void setScreen(Screen screen){
-
-        if (current_screen != null) {
-            current_screen.dispose();
-        }
-
-        current_screen = screen;
-        current_screen.create();
-    }
-
-    public static Screen getCurrentScreen(){
-        return current_screen;
-    }
-}
