+++ core/src/com/spaulding/salt/Screens/SaltScreen.java
@@ -1,0 +1,50 @@
+package com.spaulding.salt.Screens;
+
+import com.badlogic.gdx.Screen;
+import com.spaulding.salt.Salt;
+
+/**
+ * Created by connor on 5/17/16.
+ */
+public class SaltScreen implements Screen {
+    Salt game;
+
+    public SaltScreen(Salt game) {
+        this.game = game;
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
