+++ desktop/build.gradle
@@ -21,7 +21,7 @@ task dist(type: Jar) {
     from files(project.assetsDir);
  
     manifest {
-        attributes 'Main-Class': project.mainClassName
+        attributes 'Salt-Class': project.mainClassName
     }
 }
