+++ desktop/src/com/spaulding/salt/desktop/DesktopLauncher.java
@@ -2,7 +2,7 @@ package com.spaulding.salt.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.spaulding.salt.Main;
+import com.spaulding.salt.Salt;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
@@ -11,6 +11,6 @@ public class DesktopLauncher {
 		config.useGL30 = false;
 		config.width = 1080;
 		config.height = 640;
-		new LwjglApplication(new Main(), config);
+		new LwjglApplication(new Salt(), config);
 	}
 }
