+++ html/src/com/spaulding/salt/client/HtmlLauncher.java
@@ -3,7 +3,7 @@ package com.spaulding.salt.client;
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.backends.gwt.GwtApplication;
 import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
-import com.spaulding.salt.Main;
+import com.spaulding.salt.Salt;
 
 public class HtmlLauncher extends GwtApplication {
 
@@ -14,6 +14,6 @@ public class HtmlLauncher extends GwtApplication {
 
         @Override
         public ApplicationListener getApplicationListener () {
-                return new Main();
+                return new Salt();
         }
 }
